+++ app/assets/javascripts/mbeditor/application.js
@@ -1,4 +1,5 @@
 //= require mbeditor/editor_store
+//= require mbeditor/file_icon
 //= require mbeditor/file_service
 //= require mbeditor/git_service
 //= require mbeditor/search_service